requirements.txt
aiogram~=3.4.1
python-dotenv~=1.0.1
SQLAlchemy~=2.0.38
psycopg2>=2.9.0
alembic~=1.15.2
loguru~=0.7.3
